chore: move LegacyCollectorExtensions to `extension`
